+++ hw16/pkg/models/structures.go
@@ -1,0 +1,11 @@
+package pkg
+
+// Структура фильма.
+type Film struct {
+	ID        int
+	Year      int
+	Title     string
+	BoxOffice int
+	StudioID  int
+	RateID    int
+}
